[step2] refactor: domain 패키지 추가
