+++ src/main/kotlin/blackjack/domain/Card.kt
@@ -1,3 +1,3 @@
-package blackjack
+package blackjack.domain
 
 class Card(val denom: Denomination, val type: CardType)
