+++ src/main/kotlin/blackjack/domain/CardType.kt
@@ -1,4 +1,4 @@
-package blackjack
+package blackjack.domain
 
 enum class CardType {
     CLUBS, DIAMONDS, HEARTS, SPADES
